splited products table
